Modify getJobOfferCount method in JobOfferDao

+++ paw-webapp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/persistence/JobOfferDao.java
@@ -24,7 +24,7 @@ public interface JobOfferDao {
 
     List<JobOffer> getAllJobOffers();
 
-    Integer getJobOffersCount();
+    long getJobOffersCount();
 
     List<JobOffer> getJobOffersList(int page, int pageSize);
