+++ paw-webapp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/JobOfferService.java
@@ -13,7 +13,7 @@ public interface JobOfferService {
 
     List<JobOffer> getAllJobOffers();
 
-    Integer getJobOffersCount();
+    long getJobOffersCount();
 
     List<JobOffer> getJobOffersList(int page, int pageSize);
